requirements.txt
tensorflow-gpu~=1.15.0
numpy~=1.18.4
networkx~=2.4